feat: dedup while and for loop code and admit label limitation
